Change to new detector API that doesn't need diagnostic category, remove analysis
