+++ servers/Azure.Mcp.Server/docs/azmcp-commands.md
@@ -572,80 +572,18 @@ azmcp appservice webapp get --subscription "my-subscription" \
 #### Web App Diagnostics
 
 ```bash
-# Get diagnostic categories for an App Service Web App
+# List detectors for an App Service Web App
 # ❌ Destructive | ✅ Idempotent | ❌ OpenWorld | ✅ ReadOnly | ❌ Secret | ❌ LocalRequired
-azmcp appservice webapp diagnostic get-category --subscription <subscription> \
-                                                --resource-group <resource-group> \
-                                                --app <app> \
-                                                [--diagnostic-category <diagnostic-category>]
+azmcp appservice webapp diagnostic list --subscription <subscription> \
+                                        --resource-group <resource-group> \
+                                        --app <app>
 
 # Examples:
-# List diagnostic categories for an App Service Web App
+# List diagnostic detectors for an App Service Web App
 # ❌ Destructive | ✅ Idempotent | ❌ OpenWorld | ✅ ReadOnly | ❌ Secret | ❌ LocalRequired
-azmcp appservice webapp diagnostic get-category --subscription "my-subscription" \
-                                                --resource-group "my-resource-group" \
-                                                --app "my-web-app"
-
-# Get the diagnostic category for an App Service Web App
-# ❌ Destructive | ✅ Idempotent | ❌ OpenWorld | ✅ ReadOnly | ❌ Secret | ❌ LocalRequired
-azmcp appservice webapp diagnostic get-category --subscription "my-subscription" \
-                                                --resource-group "my-resource-group" \
-                                                --app "my-web-app" \
-                                                --diagnostic-category "diagnostic-category"
-```
-
-```bash
-# Get analysis for a diagnostic category
-# ❌ Destructive | ✅ Idempotent | ❌ OpenWorld | ✅ ReadOnly | ❌ Secret | ❌ LocalRequired
-azmcp appservice webapp diagnostic get-analysis --subscription <subscription> \
-                                                --resource-group <resource-group> \
-                                                --app <app> \
-                                                --diagnostic-category <diagnostic-category> \
-                                                [--analysis-name <analysis-name>]
-
-# Examples:
-# List analyses for a diagnostic category
-# ❌ Destructive | ✅ Idempotent | ❌ OpenWorld | ✅ ReadOnly | ❌ Secret | ❌ LocalRequired
-azmcp appservice webapp diagnostic get-analysis --subscription "my-subscription" \
-                                                --resource-group "my-resource-group" \
-                                                --app "my-web-app" \
-                                                --diagnostic-category "diagnostic-category"
-
-
-# Get an analysis for a diagnostic category
-# ❌ Destructive | ✅ Idempotent | ❌ OpenWorld | ✅ ReadOnly | ❌ Secret | ❌ LocalRequired
-azmcp appservice webapp diagnostic get-analysis --subscription "my-subscription" \
-                                                --resource-group "my-resource-group" \
-                                                --app "my-web-app" \
-                                                --diagnostic-category "diagnostic-category" \
-                                                --analysis-name "analysis-name"
-```
-
-```bash
-# Get detector for a diagnostic category
-# ❌ Destructive | ✅ Idempotent | ❌ OpenWorld | ✅ ReadOnly | ❌ Secret | ❌ LocalRequired
-azmcp appservice webapp diagnostic get-detector --subscription <subscription> \
-                                                --resource-group <resource-group> \
-                                                --app <app> \
-                                                --diagnostic-category <diagnostic-category> \
-                                                [--detector-name <detector-name>]
-
-# Examples:
-# List detectors for a diagnostic category
-# ❌ Destructive | ✅ Idempotent | ❌ OpenWorld | ✅ ReadOnly | ❌ Secret | ❌ LocalRequired
-azmcp appservice webapp diagnostic get-detector --subscription "my-subscription" \
-                                                --resource-group "my-resource-group" \
-                                                --app "my-web-app" \
-                                                --diagnostic-category "diagnostic-category"
-
-
-# Get a detector for a diagnostic category
-# ❌ Destructive | ✅ Idempotent | ❌ OpenWorld | ✅ ReadOnly | ❌ Secret | ❌ LocalRequired
-azmcp appservice webapp diagnostic get-detector --subscription "my-subscription" \
-                                                --resource-group "my-resource-group" \
-                                                --app "my-web-app" \
-                                                --diagnostic-category "diagnostic-category" \
-                                                --detector-name "detector-name"
+azmcp appservice webapp diagnostic list --subscription "my-subscription" \
+                                        --resource-group "my-resource-group" \
+                                        --app "my-web-app"
 ```
 
 ### Azure CLI Operations
