+++ servers/Azure.Mcp.Server/docs/e2eTestPrompts.md
@@ -108,12 +108,7 @@ This file contains prompts used for end-to-end testing to ensure each tool is in
 | appservice_webapp_get | List the web apps in my subscription |
 | appservice_webapp_get | Show me the web apps in my <resource_group> resource group |
 | appservice_webapp_get | Get the details for web app <webapp> in <resource_group> |
-| appservice_webapp_diagnostic_get-category | List the diagnostic categories for web app <webapp> in <resource_group> |
-| appservice_webapp_diagnostic_get-category | Get the diagnostic category <diagnostic_category> for web app <webapp> in <resource_group> |
-| appservice_webapp_diagnostic_get-analysis | List the analyses for diagnostic category <diagnostic_category> for web app <webapp> in <resource_group> |
-| appservice_webapp_diagnostic_get-analysis | Get the analysis <analysis_name> for diagnostic category <diagnostic_category> for web app <webapp> in <resource_group> |
-| appservice_webapp_diagnostic_get-detector | List the detectors for diagnostic category <diagnostic_category> for web app <webapp> in <resource_group> |
-| appservice_webapp_diagnostic_get-detector | Get the detector <detector_name> for diagnostic category <diagnostic_category> for web app <webapp> in <resource_group> |
+| appservice_webapp_diagnostic_list | List the detectors for diagnostic category <diagnostic_category> for web app <webapp> in <resource_group> |
 
 ## Azure Application Insights
