--- tools/Azure.Mcp.Tools.AppService/src/Options/Webapp/Diagnostic/WebappDiagnosticCategoryGetOptions.cs
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-namespace Azure.Mcp.Tools.AppService.Options.Webapp.Diagnostic;
-
-public sealed class WebappDiagnosticCategoryGetOptions : BaseWebappDiagnosticOptions;
